+++ src/fields/Products.php
@@ -255,6 +255,7 @@ class Products extends ElementField
                 'name' => 'multi',
                 'if' => 'displayType == "dropdown"',
             ]),
+            ...$this->defineElementFieldSearchableDropdownAppearanceSchema(),
             SchemaHelper::labelPosition($this),
             SchemaHelper::instructions(),
             SchemaHelper::instructionsPosition($this),