+++ src/fields/Users.php
@@ -277,6 +277,7 @@ class Users extends ElementField
                 'name' => 'multi',
                 'if' => 'displayType == "dropdown"',
             ]),
+            ...$this->defineElementFieldSearchableDropdownAppearanceSchema(),
             SchemaHelper::labelPosition($this),
             SchemaHelper::instructions(),
             SchemaHelper::instructionsPosition($this),