+++ src/fields/Variants.php
@@ -267,6 +267,7 @@ class Variants extends ElementField
                 'name' => 'multi',
                 'if' => 'displayType == "dropdown"',
             ]),
+            ...$this->defineElementFieldSearchableDropdownAppearanceSchema(),
             SchemaHelper::labelPosition($this),
             SchemaHelper::instructions(),
             SchemaHelper::instructionsPosition($this),